feat(api): set-up mail support (#35)
to make the environment more portable and easy to develop on, I have
decided we should use sendgrid. I was looking at a lot of options, like
using mailu or any custom SMTP provider.

However, those depend on DNS to deliver emails, which is not convenient
to us, so we will use this solution for the time being.